+++ core/server/apps/private-blogging/lib/helpers/index.js
@@ -1,0 +1,3 @@
+module.exports = function registerHelpers(ghost) {
+    ghost.helpers.register('input_password', require('./input_password'));
+};
