+++ core/server/apps/private-blogging/lib/helpers/input_password.js
@@ -3,11 +3,12 @@
 //
 // Password input used on private.hbs for password-protected blogs
 //
-// We use the name meta_title to match the helper for consistency:
+// We use the name input_password to match the helper for consistency:
 // jscs:disable requireCamelCaseOrUpperCaseIdentifiers
 
-var hbs             = require('express-hbs'),
-    utils           = require('./utils'),
+// Dirty requires
+var hbs = require('express-hbs'),
+    utils = require('../../../../helpers/utils'),
     input_password;
 
 input_password = function (options) {
