+++ core/server/helpers/index.js
@@ -40,8 +40,6 @@ coreHelpers.twitter_url = require('./twitter_url');
 coreHelpers.url = require('./url');
 
 // Specialist helpers for certain templates
-coreHelpers.input_password = require('./input_password');
-coreHelpers.input_email = require('./input_email');
 coreHelpers.page_url = require('./page_url');
 coreHelpers.pageUrl = require('./page_url').deprecated;
 
@@ -103,8 +101,6 @@ registerHelpers = function (adminHbs) {
     registerThemeHelper('has', coreHelpers.has);
     registerThemeHelper('is', coreHelpers.is);
     registerThemeHelper('image', coreHelpers.image);
-    registerThemeHelper('input_email', coreHelpers.input_email);
-    registerThemeHelper('input_password', coreHelpers.input_password);
     registerThemeHelper('meta_description', coreHelpers.meta_description);
     registerThemeHelper('meta_title', coreHelpers.meta_title);
     registerThemeHelper('navigation', coreHelpers.navigation);
@@ -128,7 +124,6 @@ registerHelpers = function (adminHbs) {
 
     // Register admin helpers
     registerAdminHelper('asset', coreHelpers.asset);
-    registerAdminHelper('input_password', coreHelpers.input_password);
 };
 
 module.exports = coreHelpers;
